Profile page created

+++ app/src/main/java/com/project/foodie/data/entity/Yemek.kt
@@ -1,3 +1,3 @@
 package com.project.foodie.data.entity
 
-data class Yemek(var yemekId:Int,var yemekName:String,var yemekPict:String,var yemekPrice:Double)
+data class Yemek(var yemekId:Int,var yemekName:String,var yemekPict:String,var yemekPrice:Int)
